# docs/swagger.yaml
basePath: /
definitions:
  models.Customer:
    properties:
      cidade:
        type: string
      estado:
        type: string
      grupo:
        type: string
      id:
        type: string
      id_grupo:
        type: integer
      nome:
        type: string
      nome_fantasia:
        type: string
    type: object
  models.Product:
    properties:
      descricao:
        type: string
      grupo:
        type: string
      id:
        type: string
      id_grupo:
        type: string
    type: object
host: https://techfinance.fly.dev
info:
  contact: {}
  description: TechFinance é uma solução mobile para para prover informações para
    apoio para tomada de decisões. Esta é uma API backend construída utilizando o
    framework **Fiber** (na linguagem Go).
  termsOfService: http://swagger.io/terms/
  title: TechFinance API
  version: "1.0"
paths:
  /clientes:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de clientes
      parameters:
      - description: Nome do cliente
        in: query
        name: nome
        type: string
      - description: Grupo do cliente
        in: query
        name: grupo
        type: string
      - description: Limite de registros por página
        in: query
        name: limite
        type: integer
      - description: Número da página
        in: query
        name: pagina
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Customer'
            type: array
      security:
      - BearerAuth: []
      summary: Retorna uma lista de clientes
      tags:
      - clientes
  /produtos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de produtos
      parameters:
      - description: Descrição do produto
        in: query
        name: descricao
        type: string
      - description: Grupo do produto
        in: query
        name: grupo
        type: string
      - description: Limite de registros por página
        in: query
        name: limite
        type: integer
      - description: Número da página
        in: query
        name: pagina
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      security:
      - BearerAuth: []
      summary: Retorna uma lista de produtos
      tags:
      - produtos
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"